[077/906] up
git-svn-id: svn://svn.wobow.com/GStreamer_playground/gst-plugins-gl@475 93df14bb-0f41-7a43-8087-d3e2a2f0e464

+++ gst-libs/gst/gl/gstgldisplay.c
@@ -635,7 +635,7 @@ gst_gl_display_glutCreateWindow (GstGLDisplay *display)
 
     //insert glut context to the map
     display->glutWinId = glutWinId;
-    g_hash_table_insert (gst_gl_display_map, GUINT_TO_POINTER ((guint64)glutWinId), display);
+    g_hash_table_insert (gst_gl_display_map, GUINT_TO_POINTER (glutWinId), display);
 
     //check glut id validity
     g_assert (glutGetWindow() == glutWinId);
@@ -694,7 +694,7 @@ gst_gl_display_glutDestroyWindow (GstGLDisplay *display)
 	    }
     }
 
-    g_hash_table_remove (gst_gl_display_map, GINT_TO_POINTER ((guint64)display->glutWinId)); 
+    g_hash_table_remove (gst_gl_display_map, GINT_TO_POINTER (display->glutWinId)); 
     GST_DEBUG ("glut window destroyed: %d", display->glutWinId);
 
     //if the map is empty, leaveMainloop and join the thread
@@ -868,7 +868,7 @@ gst_gl_display_checkMsgValidity (GstGLDisplayMsg *msg)
         case GST_GL_DISPLAY_ACTION_VIDEO:
         case GST_GL_DISPLAY_ACTION_REDISPLAY:
             //msg is out of date if the associated display is not in the map
-            if (!g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER ((guint64)msg->glutWinId)))
+            if (!g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER (msg->glutWinId)))
                 valid = FALSE;
             break;
         default:
@@ -1115,7 +1115,7 @@ gst_gl_display_onReshape(gint width, gint height)
   
     //retrieve the display associated to the glut context
     glutWinId = glutGetWindow ();
-    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER ((guint64)glutWinId));
+    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER (glutWinId));
 
     //glutGetWindow return 0 if no windows exists, then g_hash_table_lookup return NULL
     if (display == NULL) return;
@@ -1147,7 +1147,7 @@ void gst_gl_display_draw(void)
   
     //retrieve the display associated to the glut context
     glutWinId = glutGetWindow ();
-    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER ((guint64)glutWinId));
+    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER (glutWinId));
     
     //glutGetWindow return 0 if no windows exists, then g_hash_table_lookup return NULL
     if (display == NULL) return;
@@ -1224,7 +1224,7 @@ void gst_gl_display_onClose (void)
   
     //retrieve the display associated to the glut context
     glutWinId = glutGetWindow ();
-    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER ((guint64)glutWinId));
+    display = g_hash_table_lookup (gst_gl_display_map, GINT_TO_POINTER (glutWinId));
     
     //glutGetWindow return 0 if no windows exists, then g_hash_table_lookup return NULL
     if (display == NULL) return;
